adding changes to types etc pc269

+++ rpa-pc269/DailyGiWells.cs
@@ -9,12 +9,12 @@ namespace rpa_functions.rpa_pc269
         public int GiwellId { get; set; }
         public int DailyreportId { get; set; }
         public string WellName { get; set; }
-        public short? OnlineTime { get; set; }
+        public decimal? OnlineTime { get; set; }
         public decimal? ChokeOpening { get; set; }
-        public short? Whp { get; set; }
-        public short? Wht { get; set; }
-        public short? Bhp { get; set; }
-        public short? Bht { get; set; }
+        public decimal? Whp { get; set; }
+        public decimal? Wht { get; set; }
+        public decimal? Bhp { get; set; }
+        public decimal? Bht { get; set; }
         public decimal? GasInjectionAllocated { get; set; }
         public decimal? GasInjectionTarget { get; set; }
         public decimal? GasInjectionMeasured { get; set; }
